Optimize display of possible numbers
Includes updated layout of fixed grid and limits its usage to 9x9 and 9 as maximum number. The dynamic grid uses an adaptive text size, scaling down if the possible number do not fit into two lines.

There are still issues if a large grid and/or large possible numbers are in use.