Add layout for the main screen between 600 and 840dp, solving issues with e.g. 8' tablets.
This layout renders the keypad under the grid to assure enough space remaining for the grid.

There is still the long lasting issue that all such vertical main screen layouts are not properly resized if there is not enough space to put the keypad under the grid, e.g. if the window is nearly a square.