Stop confetti if a new game was started in a non-disturbing way
This means: If the new game was started via the FAB displaying '+', the confetti does not create new sprites, but the existing sprites continue to fall down the screen. If the game was started via the new game screen, the confetti is no longer visible when the new game begins.