Restrict minimum grid size from 2 cells to 3 cells in release flavour
This is caused by problems when calculating grids of this size: Not every variant is possible to be calculated. Thus, this very small size gets out of the way - the size 3 is fully supported and calculable in every possible variant of options.

If a user has set its grid size to 2, it will be automatically shifted to 3. The debug flavour still allows grid size 2 as they are the smallest possible grid to test with.