Add new grid calculation algorithm
This new algorithm will hopefully replace the current one in the future. Currently, the new algorithm is available if:
  - The app runs in debug flavour and you choose "New Algorithm" when creating a new game.
  - You create a non-square grid.

The algorithm works by an initial grid with given, random numbers and single cell cages. It iterates merging two small cages into a larger one if the resulting grid still has one unique solution.

The new algorithm has a few advantages, e.g. it is guaranteed to find a valid grid with a unique solution. There a few glitches which have to be solved yet: There is no difficulty rating yet, it is not yet optimized performance wise and is likely to be slower at various square variants.