speed up creating a 'pseudo grid' by avoiding to load the difficulty ratings
Background: When the user switches an option in the new game screen, a new grid gets calculated. If this calculation gets lengthy in time, a pseudo grid gets created and shown ASAP as a visual fallback. When the real grid got finished, the pseudo grid gets exchanged by the actual grid.